Separate ManagedHandlerTestHelpers into its own file
